feat: move the pull request into a full-screen route
The dock's Pulls tab couldn't hold a whole PR diff. Promote the PR to a
per-project screen in the Outlet (like Settings), entered from the footer
PR badge:

- Pulls screen with Overview (status, markdown body, merge/create actions)
  and Files changed — the PR's diff via gh pr diff, rendered with the
  Review dock's FileDiff cards, read-only
- footer PR badge navigates to /projects/:id/pulls instead of opening the
  dock
- remove the dock "Pulls" tab and the old PullsPanel

The terminal auto-hides on the subroute, same as Settings.
