+++ frontend/src/components/pulls/Pulls.tsx
@@ -1,0 +1,403 @@
+import {useState, type ReactNode} from "react"
+import {useParams} from "react-router-dom"
+import {toast} from "sonner"
+import {
+  Check,
+  ChevronDown,
+  CircleDashed,
+  Clock,
+  ExternalLink,
+  GitBranch,
+  GitMerge,
+  GitPullRequestArrow,
+  X,
+  type LucideIcon,
+} from "lucide-react"
+import {ProjectService, System, Terminal as TerminalService} from "@/lib/rpc"
+import type {ChecksRollup, PullRequestDetail} from "@/lib/api-types"
+import {useProjects} from "@/lib/projects"
+import {activeSessionId, sessionsOf} from "@/lib/sessions"
+import {useGitStatus} from "@/lib/useGitStatus"
+import {usePullRequestDetail} from "@/lib/usePullRequestDetail"
+import {cn, errorText} from "@/lib/utils"
+import {Markdown} from "@/components/Markdown"
+import {Button} from "@/components/ui/button"
+import {
+  Dialog,
+  DialogContent,
+  DialogDescription,
+  DialogFooter,
+  DialogHeader,
+  DialogTitle,
+} from "@/components/ui/dialog"
+import {
+  DropdownMenu,
+  DropdownMenuContent,
+  DropdownMenuItem,
+  DropdownMenuSeparator,
+  DropdownMenuTrigger,
+} from "@/components/ui/dropdown-menu"
+import {Input} from "@/components/ui/input"
+import {Label} from "@/components/ui/label"
+import {PullsFiles} from "./PullsFiles"
+
+// Pulls is the per-project pull-request screen: it fills the main area on top of
+// the persistent terminals (like Settings), showing the active session's branch
+// PR — status, body and the full diff — with merge, create and open actions. It
+// resolves its path from the route's project id plus the active session (the
+// exact-match useActiveSession returns empty on this subroute).
+export function Pulls() {
+  const {projectId} = useParams()
+  const {projects, sessions} = useProjects()
+  const projectPath = projects.find((p) => p.id === projectId)?.path ?? ""
+  const sessionId = projectId ? activeSessionId(sessions, projectId) : ""
+  const session = projectId
+    ? sessionsOf(sessions, projectId).find((s) => s.id === sessionId)
+    : undefined
+  const path = session?.path || projectPath
+  const status = useGitStatus(path)
+  const branch = status?.branch ?? ""
+  const {detail, loading, error, refresh} = usePullRequestDetail(path, branch)
+
+  const inject = (text: string) => {
+    if (sessionId) {
+      void TerminalService.Write(sessionId, text)
+    }
+  }
+
+  let body: ReactNode
+  if (!path) {
+    body = <Notice>No repository</Notice>
+  } else if (error) {
+    body = <Notice>Couldn’t load the pull request: {error}</Notice>
+  } else if (detail) {
+    body = <PullRequestView path={path} detail={detail} onMerged={refresh} onInject={inject}/>
+  } else if (loading) {
+    body = <Notice>Loading…</Notice>
+  } else {
+    body = <EmptyState path={path} branch={branch} onOpened={refresh}/>
+  }
+
+  return <div className="absolute inset-0 z-10 flex flex-col bg-background">{body}</div>
+}
+
+// EditState carries a pending "edit commit message" merge: which method to run
+// and the message the dialog is editing. null when the dialog is closed.
+interface EditState {
+  method: string
+  title: string
+  subject: string
+  body: string
+}
+
+interface PullRequestViewProps {
+  path: string
+  detail: PullRequestDetail
+  onMerged: () => void
+  onInject: (text: string) => void
+}
+
+function PullRequestView({path, detail, onMerged, onInject}: PullRequestViewProps) {
+  const [merging, setMerging] = useState(false)
+  const [edit, setEdit] = useState<EditState | null>(null)
+  const [tab, setTab] = useState<"overview" | "files">("overview")
+  const blocked = detail.isDraft
+    ? "Pull request is a draft"
+    : detail.mergeable === "CONFLICTING"
+      ? `Conflicts with ${detail.baseRefName}`
+      : null
+
+  const merge = async (method: string, subject = "", body = "") => {
+    setMerging(true)
+    try {
+      await ProjectService.MergePullRequest(path, method, subject, body)
+      toast.success(`Merged #${detail.number} into ${detail.baseRefName}`)
+      setEdit(null)
+      onMerged()
+    } catch (err: unknown) {
+      toast.error(`Merge failed: ${errorText(err)}`)
+    } finally {
+      setMerging(false)
+    }
+  }
+
+  const openEdit = (method: string) => {
+    setEdit(
+      method === "squash"
+        ? {
+            method,
+            title: "Squash and merge",
+            subject: `${detail.title} (#${detail.number})`,
+            body: detail.body,
+          }
+        : {
+            method,
+            title: "Create a merge commit",
+            subject: `Merge pull request #${detail.number} from ${detail.headRefName}`,
+            body: "",
+          },
+    )
+  }
+
+  return (
+    <>
+      <div className="flex-none border-b border-border px-6 pt-5">
+        <div className="flex items-start gap-4">
+          <h1 className="min-w-0 flex-1 text-lg font-semibold leading-snug">
+            <span className="text-muted-foreground">#{detail.number}</span> {detail.title}
+          </h1>
+          <div className="flex flex-none items-center gap-2">
+            <DropdownMenu>
+              <DropdownMenuTrigger
+                render={
+                  <Button size="sm" disabled={merging || blocked !== null}>
+                    <GitMerge/>
+                    {merging ? "Merging…" : "Merge"}
+                    <ChevronDown/>
+                  </Button>
+                }
+              />
+              <DropdownMenuContent align="end">
+                <DropdownMenuItem onClick={() => void merge("squash")}>
+                  Squash and merge
+                </DropdownMenuItem>
+                <DropdownMenuItem onClick={() => void merge("merge")}>
+                  Create a merge commit
+                </DropdownMenuItem>
+                <DropdownMenuItem onClick={() => void merge("rebase")}>
+                  Rebase and merge
+                </DropdownMenuItem>
+                <DropdownMenuSeparator/>
+                <DropdownMenuItem onClick={() => openEdit("squash")}>
+                  Squash and merge, edit message…
+                </DropdownMenuItem>
+                <DropdownMenuItem onClick={() => openEdit("merge")}>
+                  Create a merge commit, edit message…
+                </DropdownMenuItem>
+              </DropdownMenuContent>
+            </DropdownMenu>
+            <Button variant="ghost" size="sm" onClick={() => void System.OpenExternal(detail.url)}>
+              <ExternalLink/>
+              Open
+            </Button>
+          </div>
+        </div>
+
+        <div className="mt-3 flex flex-wrap items-center gap-x-4 gap-y-1.5 text-xs">
+          <span
+            className={cn(
+              "flex items-center gap-1.5 font-medium",
+              detail.isDraft ? "text-amber-500" : "text-emerald-500",
+            )}
+          >
+            <GitPullRequestArrow className="size-3.5"/>
+            {detail.isDraft ? "Draft" : "Open"}
+          </span>
+          <span className="flex items-center gap-1.5 font-mono text-muted-foreground">
+            <GitBranch className="size-3.5"/>
+            {detail.headRefName} → {detail.baseRefName}
+          </span>
+          <ChecksStat checks={detail.checks}/>
+          <MergeableStat mergeable={detail.mergeable} base={detail.baseRefName}/>
+        </div>
+
+        <div className="mt-4 flex gap-1">
+          <TabButton active={tab === "overview"} onClick={() => setTab("overview")}>
+            Overview
+          </TabButton>
+          <TabButton active={tab === "files"} onClick={() => setTab("files")}>
+            Files changed
+          </TabButton>
+        </div>
+      </div>
+
+      <div className="flex-1 overflow-y-auto">
+        {tab === "overview" ? (
+          <div className="mx-auto max-w-3xl px-6 py-5">
+            {detail.body.trim() !== "" ? (
+              <Markdown>{detail.body}</Markdown>
+            ) : (
+              <p className="text-sm text-muted-foreground">No description.</p>
+            )}
+          </div>
+        ) : (
+          <PullsFiles path={path} onInject={onInject}/>
+        )}
+      </div>
+
+      {edit && (
+        <MergeMessageDialog
+          edit={edit}
+          merging={merging}
+          onChange={setEdit}
+          onCancel={() => setEdit(null)}
+          onConfirm={() => void merge(edit.method, edit.subject, edit.body)}
+        />
+      )}
+    </>
+  )
+}
+
+function TabButton({active, onClick, children}: {active: boolean; onClick: () => void; children: ReactNode}) {
+  return (
+    <button
+      type="button"
+      onClick={onClick}
+      className={cn(
+        "-mb-px border-b-2 px-3 py-2 text-sm font-medium transition-colors",
+        active
+          ? "border-primary text-foreground"
+          : "border-transparent text-muted-foreground hover:text-foreground",
+      )}
+    >
+      {children}
+    </button>
+  )
+}
+
+interface MergeMessageDialogProps {
+  edit: EditState
+  merging: boolean
+  onChange: (next: EditState) => void
+  onCancel: () => void
+  onConfirm: () => void
+}
+
+function MergeMessageDialog({edit, merging, onChange, onCancel, onConfirm}: MergeMessageDialogProps) {
+  return (
+    <Dialog open onOpenChange={(next) => !next && onCancel()}>
+      <DialogContent className="sm:max-w-lg">
+        <DialogHeader>
+          <DialogTitle>{edit.title}</DialogTitle>
+          <DialogDescription>Edit the commit message, then merge.</DialogDescription>
+        </DialogHeader>
+        <div className="flex flex-col gap-3">
+          <div className="flex flex-col gap-1.5">
+            <Label htmlFor="merge-subject">Commit message</Label>
+            <Input
+              id="merge-subject"
+              value={edit.subject}
+              onChange={(e) => onChange({...edit, subject: e.target.value})}
+              autoFocus
+            />
+          </div>
+          <div className="flex flex-col gap-1.5">
+            <Label htmlFor="merge-body">Extended description</Label>
+            <textarea
+              id="merge-body"
+              value={edit.body}
+              onChange={(e) => onChange({...edit, body: e.target.value})}
+              rows={6}
+              placeholder="Optional"
+              className="min-h-24 w-full resize-y rounded-md border border-input bg-transparent px-2.5 py-1.5 text-sm shadow-xs outline-none transition-[color,box-shadow] placeholder:text-muted-foreground focus-visible:border-ring focus-visible:ring-3 focus-visible:ring-ring/50 dark:bg-input/30"
+            />
+          </div>
+        </div>
+        <DialogFooter>
+          <Button variant="ghost" onClick={onCancel}>
+            Cancel
+          </Button>
+          <Button onClick={onConfirm} disabled={merging || edit.subject.trim() === ""}>
+            <GitMerge/>
+            {merging ? "Merging…" : edit.method === "squash" ? "Squash and merge" : "Merge"}
+          </Button>
+        </DialogFooter>
+      </DialogContent>
+    </Dialog>
+  )
+}
+
+type Tone = "pass" | "fail" | "pending" | "muted"
+
+const toneClass: Record<Tone, string> = {
+  pass: "text-emerald-500",
+  fail: "text-destructive",
+  pending: "text-amber-500",
+  muted: "text-muted-foreground",
+}
+
+function Stat({icon: Icon, tone, children}: {icon: LucideIcon; tone: Tone; children: ReactNode}) {
+  return (
+    <span className={cn("flex items-center gap-1.5 font-medium", toneClass[tone])}>
+      <Icon className="size-3.5"/>
+      {children}
+    </span>
+  )
+}
+
+function ChecksStat({checks}: {checks: ChecksRollup}) {
+  const {passed, failed, pending, total} = checks
+  if (total === 0) {
+    return null
+  }
+  if (failed > 0) {
+    return <Stat icon={X} tone="fail">{failed} of {total} checks failing</Stat>
+  }
+  if (pending > 0) {
+    return <Stat icon={Clock} tone="pending">{pending} of {total} checks running</Stat>
+  }
+  return (
+    <Stat icon={Check} tone="pass">
+      {passed === 1 ? "1 check passed" : `${passed} checks passed`}
+    </Stat>
+  )
+}
+
+function MergeableStat({mergeable, base}: {mergeable: string; base: string}) {
+  if (mergeable === "CONFLICTING") {
+    return <Stat icon={X} tone="fail">Conflicts with {base}</Stat>
+  }
+  if (mergeable === "MERGEABLE") {
+    return <Stat icon={GitMerge} tone="pass">Mergeable</Stat>
+  }
+  return <Stat icon={CircleDashed} tone="muted">Checking mergeability…</Stat>
+}
+
+interface EmptyStateProps {
+  path: string
+  branch: string
+  onOpened: () => void
+}
+
+function EmptyState({path, branch, onOpened}: EmptyStateProps) {
+  const [opening, setOpening] = useState(false)
+  const openPR = async () => {
+    setOpening(true)
+    try {
+      await ProjectService.CreatePullRequest(path)
+      onOpened()
+    } catch (err: unknown) {
+      toast.error(`Couldn’t open a pull request: ${errorText(err)}`)
+    } finally {
+      setOpening(false)
+    }
+  }
+  return (
+    <div className="flex flex-1 flex-col items-center justify-center gap-4 p-8 text-center">
+      <GitPullRequestArrow className="size-8 text-muted-foreground"/>
+      <p className="text-sm text-muted-foreground">
+        {branch ? (
+          <>
+            No open pull request for{" "}
+            <span className="font-medium text-foreground">{branch}</span>.
+          </>
+        ) : (
+          "No open pull request."
+        )}
+      </p>
+      <Button variant="outline" size="sm" onClick={() => void openPR()} disabled={opening}>
+        <GitPullRequestArrow/>
+        {opening ? "Opening…" : "Open pull request"}
+        <ExternalLink/>
+      </Button>
+    </div>
+  )
+}
+
+function Notice({children}: {children: ReactNode}) {
+  return (
+    <div className="flex flex-1 items-center justify-center p-8">
+      <p className="text-sm text-muted-foreground">{children}</p>
+    </div>
+  )
+}
